remediation/requirements.txt
kafka-python==2.0.2
kubernetes==29.0.0
requests==2.31.0
python-dotenv==1.0.1
